chore(rust/test): split hmr related structs into their own files (#6445)
- Friendly to human and AI to open desired files.
