+++ crates/rolldown_binding/src/types/binding_hmr_boundary_output.rs
@@ -1,0 +1,14 @@
+use napi_derive::napi;
+
+#[napi(object)]
+#[derive(Debug)]
+pub struct BindingHmrBoundaryOutput {
+  pub boundary: String,
+  pub accepted_via: String,
+}
+
+impl From<rolldown_common::HmrBoundaryOutput> for BindingHmrBoundaryOutput {
+  fn from(value: rolldown_common::HmrBoundaryOutput) -> Self {
+    Self { boundary: value.boundary.to_string(), accepted_via: value.accepted_via.to_string() }
+  }
+}
