+++ crates/rolldown_binding/src/types/mod.rs
@@ -1,4 +1,8 @@
+pub mod binding_client_hmr_update;
+pub mod binding_generate_hmr_patch_return;
+pub mod binding_hmr_boundary_output;
 pub mod binding_hmr_output;
+pub mod binding_hmr_update;
 pub mod binding_log;
 pub mod binding_log_level;
 pub mod binding_magic_string;
