+++ crates/rolldown_common/src/hmr/client_hmr_input.rs
@@ -1,0 +1,7 @@
+use rustc_hash::FxHashSet;
+
+#[derive(Debug)]
+pub struct ClientHmrInput {
+  pub client_id: String,
+  pub executed_modules: FxHashSet<String>,
+}
