+++ crates/rolldown_common/src/hmr/client_hmr_update.rs
@@ -1,0 +1,7 @@
+use super::hmr_update::HmrUpdate;
+
+#[derive(Debug)]
+pub struct ClientHmrUpdate {
+  pub client_id: String,
+  pub update: HmrUpdate,
+}
