+++ crates/rolldown_common/src/hmr/hmr_boundary_output.rs
@@ -1,0 +1,7 @@
+use arcstr::ArcStr;
+
+#[derive(Debug)]
+pub struct HmrBoundaryOutput {
+  pub boundary: ArcStr,
+  pub accepted_via: ArcStr,
+}
