+++ crates/rolldown_common/src/hmr/mod.rs
@@ -1,2 +1,6 @@
+pub mod client_hmr_input;
+pub mod client_hmr_update;
 pub mod hmr_boundary;
-pub mod hmr_output;
+pub mod hmr_boundary_output;
+pub mod hmr_patch;
+pub mod hmr_update;
